LayoutTests/fast/frames/sandboxed-iframe-scripting-05.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../resources/js-test-pre.js"></script>
    <script>
        var allowedExecuted = 0;
        var disallowedExecuted = 0;
        window.onload = function() {
            shouldBe("disallowedExecuted", "0");
            shouldBe("allowedExecuted", "2");

            isSuccessfullyParsed();
        };
    </script>
</head>
<body>
    <iframe id="frame" src="resources/sandboxed-iframe-script-dynamic.html"></iframe>
    <script>
        description("Verify that adding a sandbox attribute at runtime does not effect the iframe's current state.");
    </script>
</body>
</html>
