third_party/blink/web_tests/platform/generic/fast/frames/sandboxed-iframe-scripting-05-expected.txt

Verify that adding a sandbox attribute at runtime does not effect the iframe's current state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS disallowedExecuted is 0
PASS allowedExecuted is 2
PASS successfullyParsed is true

TEST COMPLETE
